fix payment redirects
